Widen CallFrame.base from u16 to u32
The u16 base field overflowed at ~20,000 non-tail frames, causing a Zig
panic instead of a clean StackOverflow error. Widening to u32 supports
up to 4 billion register indices, far beyond the 65536 hard cap.

Only base is widened — dst stays u16 since it's a per-frame offset
bounded by locals_count (~255 max). The struct stays 48 bytes on 64-bit
(the u32 fills former alignment padding).

Also updates robustness tests: the "deep non-tail recursion" test now
verifies growable stack success at 10000 frames and clean StackOverflow
at 50000 frames (past the 32768 frame hard cap).

+++ src/memory.zig
@@ -529,7 +529,7 @@ pub const GC = struct {
         wind_records: []const WindRecord,
         wind_count: usize,
         dst_reg: u16,
-        dst_base: u16,
+        dst_base: u32,
     ) !Value {
         try self.maybeCollect();
 
@@ -602,7 +602,7 @@ pub const GC = struct {
         target_wind_count: usize,
         target_handler_count: usize,
         dst_reg: u16,
-        dst_base: u16,
+        dst_base: u32,
     ) !Value {
         try self.maybeCollect();
         const cont = try self.allocator.create(Continuation);
